# specs/001-backend-foundation/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Todo API - Backend Foundation
  description: |
    RESTful API for Todo Application (Milestone 1).
    Provides CRUD operations for task management.

    **Note**: Authentication and user isolation are planned for Milestone 2.
  version: 1.0.0
  contact:
    name: API Support

servers:
  - url: http://localhost:8000
    description: Local development server

paths:
  /api/tasks:
    get:
      summary: List all tasks
      description: Retrieve all tasks ordered by created_at descending (newest first).
      operationId: list_tasks
      tags:
        - Tasks
      responses:
        '200':
          description: List of tasks retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TaskRead'
              example:
                - id: 1
                  title: "Complete project documentation"
                  description: "Write README and API docs"
                  is_completed: false
                  created_at: "2026-02-04T12:00:00"
                - id: 2
                  title: "Review pull request"
                  description: null
                  is_completed: true
                  created_at: "2026-02-04T11:30:00"
        '500':
          description: Internal server error (database connection issue)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      summary: Create a new task
      description: Create a new task with title and optional description.
      operationId: create_task
      tags:
        - Tasks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate'
            example:
              title: "New task"
              description: "Optional description"
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskRead'
              example:
                id: 3
                title: "New task"
                description: "Optional description"
                is_completed: false
                created_at: "2026-02-04T14:30:00"
        '422':
          description: Validation error (missing or invalid fields)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'

  /api/tasks/{task_id}:
    parameters:
      - name: task_id
        in: path
        required: true
        description: Unique task identifier
        schema:
          type: integer
          minimum: 1
        example: 1

    get:
      summary: Get a single task
      description: Retrieve a specific task by its ID.
      operationId: get_task
      tags:
        - Tasks
      responses:
        '200':
          description: Task retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskRead'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                detail: "Task not found"

    put:
      summary: Update a task
      description: |
        Update an existing task. All fields are optional - only provided fields will be updated.
        This enables partial updates (e.g., only updating is_completed).
      operationId: update_task
      tags:
        - Tasks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
            examples:
              complete_task:
                summary: Mark task as completed
                value:
                  is_completed: true
              update_title:
                summary: Update task title
                value:
                  title: "Updated title"
              full_update:
                summary: Update all fields
                value:
                  title: "Updated title"
                  description: "Updated description"
                  is_completed: true
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskRead'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'

    delete:
      summary: Delete a task
      description: Permanently remove a task by its ID.
      operationId: delete_task
      tags:
        - Tasks
      responses:
        '200':
          description: Task deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Task deleted successfully"
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  schemas:
    TaskCreate:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
          description: Task title (required)
          example: "Complete project documentation"
        description:
          type: string
          maxLength: 2000
          nullable: true
          description: Optional task description
          example: "Write comprehensive README and API documentation"

    TaskUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
          nullable: true
          description: Updated task title
        description:
          type: string
          maxLength: 2000
          nullable: true
          description: Updated task description
        is_completed:
          type: boolean
          nullable: true
          description: Updated completion status

    TaskRead:
      type: object
      required:
        - id
        - title
        - is_completed
        - created_at
      properties:
        id:
          type: integer
          description: Unique task identifier
          example: 1
        title:
          type: string
          description: Task title
          example: "Complete project documentation"
        description:
          type: string
          nullable: true
          description: Task description
          example: "Write comprehensive README"
        is_completed:
          type: boolean
          description: Whether the task is completed
          example: false
        created_at:
          type: string
          format: date-time
          description: Task creation timestamp (UTC)
          example: "2026-02-04T12:00:00"

    Error:
      type: object
      required:
        - detail
      properties:
        detail:
          type: string
          description: Error message
          example: "Task not found"

    ValidationError:
      type: object
      required:
        - detail
      properties:
        detail:
          type: array
          items:
            type: object
            properties:
              type:
                type: string
                example: "string_too_long"
              loc:
                type: array
                items:
                  oneOf:
                    - type: string
                    - type: integer
                example: ["body", "title"]
              msg:
                type: string
                example: "String should have at most 255 characters"
              input:
                type: string
              ctx:
                type: object

tags:
  - name: Tasks
    description: Task management operations (CRUD)